# mini-crm/docs/openapi.yaml
openapi: 3.0.0
info:
  title: MiniCRM API
  version: 1.0.0
  description: CRM management API for customers, products, and orders.
servers:
  - url: http://localhost:3000/api
...
paths:
  /customers:
    get:
      tags:
        - Customers
      summary: List all customers
      responses:
        '200':
          description: OK
    post:
      tags:
        - Customers
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                phone:
                  type: string
                email:
                  type: string
                address:
                  type: string
                notes:
                  type: string
      responses:
        '201':
          description: Created
...
  /products:
    get:
      tags:
        - Products
      summary: List all products
      responses:
        '200':
          description: OK
    post:
      tags:
        - Products
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                sku:
                  type: string
                price:
                  type: number
                  format: float
                stockQuantity:
                  type: integer
                isStockTracked:
                  type: boolean
                  default: true
      responses:
        '201':
          description: Created
...
  /orders:
    post:
      tags:
        - Orders
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customerId:
                  type: integer
                shippingAddress:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: integer
                      quantity:
                        type: integer
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request (Insufficient Stock)
...